Initial Commit
Previous changes were done offline.  Since this is becoming a thing,
moving to github.

If some errors persist, it may be because I used "published" version
when making my changes and then reintegrated into this file by hand.
I'm not terribly familiar by github, so I'm sure there's a better way to
do it, but alas, I don't know it.